Get taxa list from the Biological Database of South Australia (BDBSA)

Get taxa list from the Biological Database of South Australia (BDBSA)

get_bdbsa_taxa(
  source = "web",
  url =
    c("https://data.environment.sa.gov.au/Content/Publications/fauna-bdbsa-taxonomy.xlsx",
    "https://data.environment.sa.gov.au/Content/Publications/vascular-plants-bdbsa-taxonomy.xlsx",
    "https://data.environment.sa.gov.au/Content/Publications/other-taxonomic-groups-bdbsa-taxonomy.xlsx"),
  out_dir = tempdir(),
  standardise_taxonomy = FALSE,
  taxonomy = NULL,
  taxonomy_file = tempfile(),
  needed_ranks = c("species", "subspecies"),
  overrides = NULL,
  taxonomic_groups = c("fauna", "vascular-plants", "other"),
  filt_ranks = NULL,
  return_scientific = FALSE,
  current = TRUE
)

Arguments

source

Character. Where to source the BDBSA taxonomy. Either 'web' (default) or 'oracle' (for internal DEW use only).

url

If source = "web", a character vector of url paths to the relevant taxonomy xlsx links on https://data.environment.sa.gov.au.

out_dir

Output directory for downloaded taxonomy xlsx files if source == "web".

standardise_taxonomy

Logical. Standardise the taxa to the ALA taxonomic backbone via envClean::make_taxonomy?

taxonomy

If !standardise_taxonomy, an optional alternative existing taxonomy list containing lutaxa and taxonomy elements, as produced by envClean::make_taxonomy.

taxonomy_file

If standardise_taxonomy, file path to save taxonomy results from envClean::make_taxonomy.

needed_ranks

If standardise_taxonomy, character vector of taxonomic ranks required in the returned list, as per envClean::make_taxonomy. Can be "all" or any combination of ranks from envClean::lurank greater than or equal to subspecies. Each rank receives its own taxonomic list with taxa binned up to that level, e.g. if "genus" then all species and subspecies are binned to genus, but also returns all other taxa at levels higher than genus, i.e. family through to kingdom.

overrides

Dataframe of taxa to override, as per envClean::make_taxonomy.

taxonomic_groups

Optional character vector of broad taxonomic groups to filter the results, e.g. if only want fauna, use taxonomic_groups = "fauna". Options are 'fauna', 'vascular-plants' & 'other' (for non-fauna and non-vascular plants).

filt_ranks

If standardise_taxonomy, a character vector of taxonomic ranks to keep. Note, these may be different to needed_ranks, which set the lowest taxonomic level to bin to, e.g. with needed_ranks = "species" but without any filt_ranks this would return all taxa >= "species", but with filt_ranks = "species" would only return species level taxa.

return_scientific

Logcial. Return the scientific_name as per the ALA taxonomic backbone (e.g. as opposed to species name in ALA which can be different)?

current

Logical. Return only current BDBSA taxonomy? Applicable to source = "oracle" only.

Value

Dataframe with BDBSA original names, common names, NSX codes, indigenous status ('ind' field) and standardised taxa names from ALA (if standardise_taxonomy or taxonomy is used).