Make a region (area) of interest

Make a region (area) of interest

make_aoi(
  polygons,
  filt_col = NULL,
  filt_level = NULL,
  is_regex = FALSE,
  dissolve = TRUE,
  buffer = FALSE,
  bbox = FALSE,
  clip = NULL,
  clip_buf = 100,
  buf_crs = 7845,
  out_crs = 4326,
  densify = NULL
)

Arguments

polygons

sf

filt_col

Character name of column in polygons to filter on (or NULL for no filter)

filt_level

Character. Level(s) of filt_col to filter

is_regex

Logical. Use filt_level as a regex?

dissolve

Logical. Dissolve polygons via dplyr::summarise()?

buffer

Numeric (or FALSE). Distance to buffer in units of buf_crs

bbox

Logical. Return a bounding box around the result?

clip

sf. If not null, this will be used to clip back the original polygons +/- filtered +/- buffered +/- bbox. CHECK both CRS and clip_buf

clip_buf

Numeric (or FALSE). Distance to buffer clip in units of clip

buf_crs

EPSG code to use when buffering

out_crs

EPSG code for coordinate system of result

densify

Numeric. interval argument to terra::densify() which is used before any polygon is transformed between coordinate reference systems.