Generate a tibble of epochs.

Generate a tibble of epochs.

make_epochs(
  min_year = 1987,
  max_year = 2020,
  epoch_step = 5,
  epoch_overlap = TRUE
)

Arguments

min_year
Minimum year in epochs

max_year
Maximum year in epochs

epoch_step
Numeric. How many years in an epoch?

epoch_overlap
Logical. Should epochs overlap by one year? i.e. epoch_step = 10 with epoch_overlap = TRUE gives, say, 2000-2010 and 2010-2020 whereas epoch_overlap = FALSE gives 2000-2009 and 2010-2019.

Value

Tibble with columns

start

First year of epoch

end

Last year of epoch

years

List column containing the years for the epoch

epoch

Last two digits of first year and last year in epoch. e.g. 10-19.

data

List column of full path of raster files to summarise.

Examples

now <- as.numeric(format(Sys.Date(), "%Y"))
make_epochs(now - 30, now, 10)
#> # A tibble: 3 x 5
#>   start   end years      epoch epoch_now
#>   <dbl> <dbl> <list>     <ord> <lgl>    
#> 1  1993  2003 <int [11]> 93-03 FALSE    
#> 2  2003  2013 <int [11]> 03-13 FALSE    
#> 3  2013  2023 <int [11]> 13-23 TRUE     
make_epochs(now - 30, now, 10, F)
#> # A tibble: 4 x 5
#>   start   end years      epoch epoch_now
#>   <dbl> <dbl> <list>     <ord> <lgl>    
#> 1  1993  1993 <int [1]>  93-93 FALSE    
#> 2  1994  2003 <int [10]> 94-03 FALSE    
#> 3  2004  2013 <int [10]> 04-13 FALSE    
#> 4  2014  2023 <int [10]> 14-23 TRUE