Make a list of data frames for months and seasons

Make a list of data frames for months and seasons

make_seasons(
  start_year,
  end_year,
  seasons = NULL,
  dec_adjust = 1,
  include_all = TRUE
)

Arguments

start_year, end_year

Numeric. Year of first and last season

seasons

Dataframe mapping each month to a season. Default (NULL) will use south hemisphere seasons. Needs columns mon (numeric 1:12), month (character name to use for months) and season (character name of season to use for that month)`

dec_adjust

Logical. Adjust the year in which December is used? If TRUE (default) December will be included in the following year's first season (i.e. December 1999 will be included in, say, summer 2000)

include_all

Logical. Include a 'season' that is 'all' (and starts at calendar year).

Value

List with elements itemluseasonsDataframe. Lookup used to translate months to seasons. Equal to seasons if that was provided by seasons argument itemmonthsDataframe for each month from start_year to end_year, including start and end dates itemseasonsDataframe for each season from start_year to end_year, including start and end dates