Create output directory paths or parse output directory paths

Either create file path for saving outputs or parse the meta data from a path

name_env_out_new(
  set_list,
  base_dir = NULL,
  all_files = FALSE,
  show_null = FALSE
)

Arguments

set_list

List (optionally nested) of critical settings (only) for use in output path names. The first elements of the list will return directories. Any elements within each of the first elements of the list are concatenated to form the name of each directory (see examples)

base_dir

Character. Directory prepended to the output path

all_files

Logical or numeric. Return files within the path column provided in the resulting dataframe? If numeric, passed to the recurse argument of fs::dir_ls().

show_null

Logical. Display "NULL" or "NA" in names (or gsub it out with "")

Value

Tibble containing:

all contexts: named as per the second level elements of set_list

directories: a directory for each element in the first level of set_list

path: the output path, prepended with base_dir

Examples

settings <- list(extent = list(polygons = "sa_ibrasub_xn"
                               , filt_col = NULL
                               , filt_level = NULL
                               , bbox = TRUE
                               , buffer = 0
                               , temp_ext = "P50Y"
                               )
                 , grain = list(x = 90
                                , y = 90
                                , z = "P50Y"
                                , taxonomic = "species"
                                )
                 , reliability = list(rel_geo = 10000
                                      , rel_temp = "P10Y"
                                      )
                 )

settings$out_dir <- name_env_out(set_list = settings#[1:3]
                                 , base_dir = here::here("inst", "examples")
                                 )$path
#> Error in name_env_out(set_list = settings, base_dir = here::here("inst",     "examples")): argument "x" is missing, with no default

settings$out_dir
#> NULL

fs::dir_create(settings$out_dir)

# create some junk files
purrr::map(1:10
           , \(x) tempfile(tmpdir = settings$out_dir)
           ) %>%
  fs::file_create()
#> Error in purrr::map(1:10, function(x) tempfile(tmpdir = settings$out_dir)): ℹ In index: 1.
#> Caused by error in `tempfile()`:
#> ! invalid 'tempdir' value

files <- name_env_out(set_list = settings[1:3]
                      , base_dir = here::here("inst", "examples")
                      , all_files = TRUE
                      ) %>%
  tidyr::unnest(cols = c(files)
                , keep_empty = TRUE
                )
#> Error in name_env_out(set_list = settings[1:3], base_dir = here::here("inst",     "examples"), all_files = TRUE): argument "x" is missing, with no default

files
#> Error in eval(expr, envir, enclos): object 'files' not found